nvcc: improve formatting of -code/-arch/-gencode
This includes using the short form in text because it takes up much less
space and keeps the focus on the content.

I also tried to split the text into paragraphs where it seemed to make
sense, but I was cautious in doing so, so large parts are still block
text. Longer commands have been extracted into verbatim blocks so that
they stand out more.